Refactored tests for meta_lb , split it into multiple modules for cleaner read
